# PJM/contracts/openapi/projectmind-api.v1.json
{
  "components": {
    "schemas": {
...
      "CreateEvaluationRequest": {
        "additionalProperties": false,
        "description": "Result を変更せず追加する人工評価 request。",
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 4000,
            "title": "Comment",
            "type": "string"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRevisionRequest"
            },
            "maxItems": 100,
            "title": "Revisions",
            "type": "array"
          },
          "verdict": {
            "$ref": "#/components/schemas/EvaluationVerdict"
          }
        },
        "required": [
          "rating",
          "verdict"
        ],
        "title": "CreateEvaluationRequest",
        "type": "object"
      },
...
      "EvaluationHistoryResponse": {
        "description": "Run Result に追加された Evaluation 履歴 response。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EvaluationHistoryResponse",
        "type": "object"
      },
      "EvaluationResponse": {
        "description": "一件の追加式 Evaluation の公開 response。",
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluation_id": {
            "format": "uuid",
            "title": "Evaluation Id",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "result_id": {
            "format": "uuid",
            "title": "Result Id",
            "type": "string"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRevisionResponse"
            },
            "title": "Revisions",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/EvaluationVerdict"
          }
        },
        "required": [
          "evaluation_id",
          "result_id",
          "run_id",
          "user_id",
          "rating",
          "verdict",
          "comment",
          "revisions",
          "created_at"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
      "EvaluationRevisionRequest": {
        "additionalProperties": false,
        "description": "Result field に対する一つの JSON Pointer 修正提案。",
        "properties": {
          "pointer": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "^/",
            "title": "Pointer",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "suggested_value": {
            "title": "Suggested Value"
          }
        },
        "required": [
          "pointer",
          "suggested_value",
          "reason"
        ],
        "title": "EvaluationRevisionRequest",
        "type": "object"
      },
      "EvaluationRevisionResponse": {
        "description": "AI 原値と人工提案値を並列表示する revision response。",
        "properties": {
          "original_value": {
            "title": "Original Value"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "suggested_value": {
            "title": "Suggested Value"
          }
        },
        "required": [
          "pointer",
          "original_value",
          "suggested_value",
          "reason"
        ],
        "title": "EvaluationRevisionResponse",
        "type": "object"
      },
      "EvaluationVerdict": {
        "description": "人工評価で選択できる Result の正確性判定。",
        "enum": [
          "accurate",
          "partially_accurate",
          "inaccurate",
          "uncertain"
        ],
        "title": "EvaluationVerdict",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ProjectMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/runs/{run_id}/evaluations": {
      "get": {
        "description": "Project access 後に Result の Evaluation 履歴を返す。",
        "operationId": "list_evaluations_api_v1_projects__project_id__runs__run_id__evaluations_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evaluations",
        "tags": [
          "evaluations"
        ]
      },
      "post": {
        "description": "評価者の実 user identity を保持して Project-scoped Evaluation を追加する。",
        "operationId": "create_evaluation_api_v1_projects__project_id__runs__run_id__evaluations_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid revision"
          },
          "404": {
            "description": "Run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
...
  }
}